# 7 more algorithm optimization