--- a/cli/Cargo.toml
+++ b/cli/Cargo.toml
@@ -16,3 +16,4 @@ tokio = { version = "1.36.0", features = ["full"] }
 dotenv = "0.15.0"
 dialoguer = "0.11.0"
 console = "0.15.8"
+log = "0.4.21"