--- a/cli/src/main.rs
+++ b/cli/src/main.rs
@@ -4,7 +4,7 @@ use db::Client;
 use dialoguer::{theme::ColorfulTheme, Select};
 use std::{fmt::Debug, path::Path};
 
-use log::{Record, Level, Metadata};
+use log::{Level, Metadata, Record};
 
 struct SimpleLogger;
 
@@ -156,7 +156,9 @@ async fn main() -> Result<()> {
     }
 
     log::set_logger(&LOGGER)
-        .map(|()| log::set_max_level(log::LevelFilter::Info)).map_err(|e| anyhow!(e)).context("could not set up logger")?;
+        .map(|()| log::set_max_level(log::LevelFilter::Info))
+        .map_err(|e| anyhow!(e))
+        .context("could not set up logger")?;
 
     let args = Args::parse();