| OS      | Spec                             | Benchmark              | Before                          |                After                |
|---------|----------------------------------|------------------------|---------------------------------|:-----------------------------------:|
| Windows | 16 Cores 16 GB RAM (high stress) | 8 teams 27 time slots  | 1.053s (54K iters)              |          1.612s (55k iters)         |
| Windows | 16 Cores 16 GB RAM (high stress) | 16 teams 27 time slots | 33.948s (67K iters 66% correct) | 26.53796s (100K iters 100% correct) |
| Windows | 16 Cores 16 DB RAM (low stress)  | 16 teams 27 time slots | 19.906s (67K iters 55% correct) | 20.593s (100K iters 100% correct)   |
| Ubuntu  | 2 Cores 6 GB Ram                 | 8 teams 27 time slots  | 321.557s                        |                41.5s                |
| Ubuntu  | 4 Cores 4 GB Ram                 | 12 teams 27 time slots | 360.2s                          |              41.37554s              |